Car Rental Website/requirements.txt
alembic==1.7.4
backcall==0.2.0
certifi==2021.10.8
click==8.0.3
colorama==0.4.4
dataclasses==0.6
decorator==5.1.0
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.2
sklearn==1.1.1
idna==3.3
importlib-metadata==4.8.1
importlib-resources==5.3.0
# ipython==7.30.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.2
Mako==1.1.5
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
parso==0.8.2
pickleshare==0.7.5
pip==21.0.1
prompt-toolkit==3.0.23
Pygments==2.10.0
PyMySQL==1.0.2
python-decouple==3.4
PyYAML==6.0
setuptools==58.0.4
SQLAlchemy==1.4.26
# traitlets==5.1.1
wcwidth==0.2.5
Werkzeug==2.0.2
wheel==0.37.0
wincertstore==0.2
WTForms==2.3.3
zipp==3.6.0
